Introduce utils module
