+++ lib/clowne/planner.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'clowne/plan'
+require 'clowne/utils/plan'
 
 module Clowne
   class Planner # :nodoc: all
@@ -11,7 +11,7 @@ module Clowne
 
         declarations += compile_traits(cloner, traits) unless traits.nil?
 
-        declarations.each_with_object(Plan.new(cloner.adapter.registry)) do |declaration, plan|
+        declarations.each_with_object(Utils::Plan.new(cloner.adapter.registry)) do |declaration, plan|
           declaration.compile(plan)
         end
       end
