+++ spec/clowne/utils/params_spec.rb
@@ -1,4 +1,4 @@
-describe Clowne::Params do
+describe Clowne::Utils::Params do
   describe 'build proxy' do
     let(:params) do
       {
@@ -13,7 +13,7 @@ describe Clowne::Params do
     context 'when value is true' do
       let(:value) { true }
 
-      it { is_expected.to be_a(Clowne::Params::PassProxy) }
+      it { is_expected.to be_a(Clowne::Utils::Params::PassProxy) }
 
       it 'return all params' do
         expect(permitted_params).to eq(params)
@@ -23,7 +23,7 @@ describe Clowne::Params do
     context 'when value is false' do
       let(:value) { false }
 
-      it { is_expected.to be_a(Clowne::Params::NullProxy) }
+      it { is_expected.to be_a(Clowne::Utils::Params::NullProxy) }
 
       it 'return empty hash' do
         expect(permitted_params).to eq({})
@@ -33,7 +33,7 @@ describe Clowne::Params do
     context 'when value is false' do
       let(:value) { nil }
 
-      it { is_expected.to be_a(Clowne::Params::NullProxy) }
+      it { is_expected.to be_a(Clowne::Utils::Params::NullProxy) }
 
       it 'return empty hash' do
         expect(permitted_params).to eq({})
@@ -47,7 +47,7 @@ describe Clowne::Params do
       context 'with 1 args' do
         let(:value) { ->(p) { p[:profile][:data] } }
 
-        it { is_expected.to be_a(Clowne::Params::BlockProxy) }
+        it { is_expected.to be_a(Clowne::Utils::Params::BlockProxy) }
 
         it 'return nested hash' do
           expect(permitted_params).to eq(name: 'Robin')
@@ -57,7 +57,7 @@ describe Clowne::Params do
       context 'with 2 args' do
         let(:value) { ->(p, parent) { p[:profile][:data].merge(parent: parent) } }
 
-        it { is_expected.to be_a(Clowne::Params::BlockProxy) }
+        it { is_expected.to be_a(Clowne::Utils::Params::BlockProxy) }
 
         it 'return nested hash with record' do
           expect(permitted_params).to eq(name: 'Robin', parent: parent)
@@ -68,7 +68,7 @@ describe Clowne::Params do
     context 'when value is a key' do
       let(:value) { :profile }
 
-      it { is_expected.to be_a(Clowne::Params::KeyProxy) }
+      it { is_expected.to be_a(Clowne::Utils::Params::KeyProxy) }
 
       it 'return nested hash' do
         expect(permitted_params).to eq(data: { name: 'Robin' })
@@ -78,7 +78,7 @@ describe Clowne::Params do
     context 'when value is a key but nested values is not a Hash' do
       let(:value) { :rating }
 
-      it { is_expected.to be_a(Clowne::Params::KeyProxy) }
+      it { is_expected.to be_a(Clowne::Utils::Params::KeyProxy) }
 
       it 'raise KeyError' do
         expect { permitted_params }.to raise_error(KeyError, "value by key 'rating' must be a Hash")
